HIC: split off the lexer into an extra class
